old/requirements.txt
opencv-python
numpy
windows_filedialogs